requirements.txt
AfricastalkingGateway==1.9
celery==4.1.0
google-api-python-client==1.6.4
httplib2==0.10.3
oauth2client==4.1.2
